modularize layer manager
