+++ pykrita/comfyui/extension/layer_manager/__init__.py
@@ -1,0 +1,2 @@
+from .manager import LayerManager  # noqa: F401
+from .persistent_manager import PersistentLayerManager  # noqa: F401
