+++ pykrita/comfyui/extension/ui/docker/nodes/comfyui_node.py
@@ -2,7 +2,7 @@ from typing import ClassVar
 
 from PyQt5.QtWidgets import QWidget
 
-from ....layers_manager import PersistentLayerManager
+from ....layer_manager import LayerManager
 from ....models import Node, NodeDirection
 
 
@@ -10,9 +10,9 @@ class ComfyUiNode(QWidget):
     type: ClassVar[str | None] = None
     direction: NodeDirection
 
-    def __init__(self, node: Node, layers_manager: PersistentLayerManager) -> None:
+    def __init__(self, node: Node, layer_manager: LayerManager) -> None:
         super().__init__()
         self.direction = node.direction
-        self.layers_manager = layers_manager
+        self.layer_manager = layer_manager
 
     def cleanup(self) -> None: ...
